重构设置页
