+++ src/options/about/index.js
@@ -1,0 +1,5 @@
+import template from './about.html';
+
+export default {
+  template
+};
