--- src/options/templates/index.js
@@ -1,5 +1,0 @@
-import settings from './settings.html';
-import about from './about.html';
-import voices from './voices.html';
-
-export {settings,about,voices};
